+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Fix bind and make any addresses
 - Add expected host method to address
 - Deprecate old and unused functions
-- Make future migration to std::format easier
+- Supports C++20 and later std::format
 
 ## v0.9.8
 - Fix event sync